Rename MetaDataSchemaContainer to MetaDataSchemaCollection
